Add buildroot image and workflow
In order to build shim reproducibly, the build environment needs to be
reproducible. Unlike Debian's snapshot service, Endless does have access
to old packages. The container image here is intended to contain all
necessary build dependencies so that shim can be built reproducibly in
it. The image should be built and published whenever shim is built for
Endless OS so that it matches that build environment as well as
possible.